mock-generator/pyproject.toml
[project]
name = "mock-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.1.2",
    "kafka-python-ng>=2.2.3",
    "python-decouple>=3.8",
    "redis>=7.1.0",
]
